Read ROM error
This change solve the problem that when we read the ROM, the memory array set to the same lenght to the lengh of the ROM file.
If the program try to read some memory location upper than the length of the memory array (same length than ROM file), emulator crashed.
Example: memory has 4096 positions (bytes), if ROM file had 373 bytes to read, emulator set the memory as 373 as max lenght. If emulator tried to read the position in the memory 400, emulator crash (out of index, obviously).